+++ lib/puppet/type/wls_domain_partition/domain_partition_name.rb
@@ -1,0 +1,9 @@
+newparam(:domain_partition_name) do
+  include EasyType
+  include EasyType::Validators::Name
+
+  isnamevar
+
+  desc 'The domain partition group name'
+
+end